Create a separate smoke test suite for Spring Scheudler
