+++ test/smoke/testApps/SpringScheduler/build.gradle
@@ -1,0 +1,21 @@
+apply plugin: 'war'
+
+sourceCompatibility = 1.8
+targetCompatibility = 1.8
+compileSmokeTestJava.sourceCompatibility = 1.8
+compileSmokeTestJava.targetCompatibility = 1.8
+
+ext.testAppArtifactDir = war.destinationDirectory
+ext.testAppArtifactFilename = war.archiveFileName.get()
+
+dependencies {
+    compileOnly aiAgentJar // this is just to force building the agent first
+
+    compile aiWebAutoJar
+    compile (group: 'org.springframework.boot', name: 'spring-boot-starter-web', version: '2.1.7.RELEASE') {
+        exclude group: 'org.springframework.boot', module: 'spring-boot-starter-tomcat'
+    }
+    compile group: 'org.reactivestreams', name: 'reactive-streams', version: '1.0.3'
+
+    providedCompile 'javax.servlet:javax.servlet-api:3.0.1'
+}
