+++ test/smoke/testApps/SpringScheduler/src/main/java/com/microsoft/ajl/simple/SpringSchedulerApp.java
@@ -1,0 +1,34 @@
+package com.microsoft.ajl.simple;
+
+import java.io.IOException;
+import java.io.InputStream;
+import java.net.HttpURLConnection;
+import java.net.URL;
+
+import com.microsoft.applicationinsights.core.dependencies.google.common.io.ByteStreams;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.builder.SpringApplicationBuilder;
+import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
+import org.springframework.scheduling.annotation.EnableScheduling;
+import org.springframework.scheduling.annotation.Scheduled;
+
+@SpringBootApplication
+@EnableScheduling
+public class SpringSchedulerApp extends SpringBootServletInitializer {
+
+    @Override
+    protected SpringApplicationBuilder configure(SpringApplicationBuilder applicationBuilder) {
+        return applicationBuilder.sources(SpringSchedulerApp.class);
+    }
+
+    @Scheduled(fixedRate = 100)
+    public int fixedRateScheduler() throws IOException {
+        URL obj = new URL("https://www.bing.com/search?q=spaces%20test");
+        HttpURLConnection connection = (HttpURLConnection) obj.openConnection();
+        connection.getContentType();
+        InputStream content = connection.getInputStream();
+        ByteStreams.exhaust(content);
+        content.close();
+        return connection.getResponseCode();
+    }
+}
