+++ test/smoke/testApps/SpringScheduler/src/smokeTest/resources/appServers.txt
@@ -1,0 +1,1 @@
+tomcat85
